nucliadb/requirements.txt
uvicorn
pydantic_argparse
nucliadb-one>=1.0.1
nucliadb-node-binding>=0.2.0


aiohttp==3.8.1
clandestined==1.0.1
tikv-client>=0.0.3
lru-dict==1.1.7
backoff
aiofiles==0.8.0
types-aiofiles==0.8.3
protobuf==3.20.2
types-protobuf>=3.19.20
grpcio>=1.44.0
grpcio-health-checking>=1.44.0
grpcio-channelz>=1.44.0
grpcio-status>=1.44.0
grpcio-tools>=1.44.0
grpcio-testing>=1.44.0
grpcio-reflection>=1.44.0
orjson==3.6.7
aiomonitor
types-setuptools
pydantic
sentry-sdk>=1.5.12
prometheus-client==0.12.0
opentelemetry-instrumentation-aiohttp-client==0.30b1
aioredis==2.0.1
aiobotocore==2.1.1
botocore
google-cloud-storage
gcloud
oauth2client
mmh3==3.0.0
tikv-client==0.0.3
nucliadb-protos>=0.0.29
nucliadb-models>=1.0.1

pydantic==1.9.0
protobuf
nucliadb-protos>=0.0.29

starlette-prometheus==0.9.0
uvicorn
opentelemetry-instrumentation-fastapi==0.30b1
opentelemetry-instrumentation-aiohttp-client
opentelemetry-propagator-b3
fastapi-versioning==0.10.0
fastapi==0.75.0
pydantic==1.9.0
sentry-sdk>=1.5.12
pyjwt==2.4.0
mmh3==3.0.0
httpx==0.23.0
aiofiles==0.8.0
types-pkg-resources==0.1.3
grpc-stubs==1.24.7
orjson==3.6.7
opentelemetry-semantic-conventions==0.30b1
aiodns==3.0.0
nucliadb-protos>=0.0.29
nucliadb-utils[storages,fastapi,cache]>=1.0.9
nucliadb-ingest>=1.0.0
nucliadb-models>=1.0.0
nucliadb-telemetry>=1.1.6

starlette-prometheus==0.9.0
uvicorn==0.16.0
opentelemetry-instrumentation-fastapi==0.30b1
opentelemetry-instrumentation-aiohttp-client
opentelemetry-propagator-b3
fastapi-versioning==0.10.0
fastapi==0.75.0
pydantic==1.9.0
sentry-sdk>=1.5.12
pyjwt==2.4.0
mmh3==3.0.0
httpx==0.23.0
types-orjson
types-aiofiles
types-setuptools
aiohttp==3.8.1
nucliadb-protos>=0.0.29
nucliadb-utils[storages,fastapi,cache]>=1.0.9
nucliadb-ingest>=1.0.0
nucliadb-models>=1.0.0
nucliadb-telemetry>=1.1.6
